Refactor `test_parse_expression.cpp`

+++ include/lexer.h
@@ -17,7 +17,7 @@ enum class Token {
 TokValPair* get_token(const char*&);
 TokValPair* peek(const char* iterator);
 
-int string_to_int(const std::string, int&);
+int string_to_int(const std::string, long&);
 int string_to_double(const std::string, double&);
 
 #endif
