+++ include/t_vals.h
@@ -6,7 +6,7 @@
 struct TVals {
   std::string num_str;
   std::string ident_str;
-  int int_num_val;
+  long int_num_val;
   double double_num_val;
 };
 #endif
